refactor: Create elements folder containing
- Contains all of the element related files
