+++ plugins/ui/src/js/src/elements/IconElementUtils.ts
